修改

+++ Server/xiaomi_mall/src/main/java/com/xiaomi_mall/common/BaseContext.java
@@ -1,13 +1,13 @@
 package com.xiaomi_mall.common;
 
 public class BaseContext {
-    private static ThreadLocal<Long> threadLocal = new ThreadLocal<>();
+    private static ThreadLocal<Integer> threadLocal = new ThreadLocal<>();
 
-    public static void setCurrentId(Long id) {
+    public static void setCurrentId(Integer id) {
         threadLocal.set(id);
     }
 
-    public static Long getCurrentId() {
+    public static Integer getCurrentId() {
         return threadLocal.get();
     }
 }
